refactor(content client): reuse client
