+++ src/lib/client/createBaseContentClient.ts
@@ -1,0 +1,24 @@
+import { CommonContentClientConfig } from '../config/CommonContentClientConfig';
+import Axios, { AxiosInstance } from 'axios';
+
+/**
+ * Create base network client to make requests to the content delivery service
+ * @param config
+ * @hidden
+ */
+export function createBaseContentClient(
+  config: CommonContentClientConfig,
+  defaultHost
+): AxiosInstance {
+  const client = Axios.create({
+    adapter: config.adaptor,
+    timeout: config.timeout || 0,
+  });
+
+  if (config.stagingEnvironment) {
+    client.defaults.baseURL = `https://${config.stagingEnvironment}`;
+  } else {
+    client.defaults.baseURL = config.baseUrl || defaultHost;
+  }
+  return client;
+}
