+++ src/lib/content/coordinators/FilterBy.spec.ts
@@ -7,13 +7,16 @@ import * as PAGED_RESPONSE from './__fixtures__/filterBy/PAGED_RESPONSE.json';
 
 import { FilterBy } from './FilterBy';
 import { HttpError } from '../model/HttpError';
+import { createContentClient } from '../../client/createContentClient';
 
 use(chaiAsPromised);
 
 function createCoordinator<T = any>(config: any): [MockAdapter, FilterBy<T>] {
   const mocks = new MockAdapter(null);
-  const client = new FilterBy<T>({ ...config, adaptor: mocks.adapter() });
-  return [mocks, client];
+  const mergedConfig = { ...config, adaptor: mocks.adapter() };
+  const client = createContentClient(mergedConfig);
+  const coordinator = new FilterBy<T>(mergedConfig, client);
+  return [mocks, coordinator];
 }
 
 const cd2RunConfig = {
